+++ scripts/templates/template_conf_load_and_train.lua
@@ -1,4 +1,7 @@
 #!/usr/bin/env lua
+local Help = dofile("scripts/modules/help_cli.lua")
+Help.auto_exit_help()
+
 -- scripts/templates/template_conf_load_and_train.lua
 --
 -- Template: charger une config via --conf et entraîner un modèle